+++ tests/test_cmd_process.py
@@ -7,6 +7,24 @@ from tm1cli.main import app
 runner = CliRunner()
 
 
+@pytest.mark.parametrize("command", ["list", "ls"])
+@pytest.mark.parametrize(
+    "raw_option,expected_output",
+    [
+        (None, "- Process1\n- Process2\n"),
+        ("--output-raw", "Process1\nProcess2\n"),
+    ],
+)
+def test_process_list(mocker, command, raw_option, expected_output):
+    mocker.patch("tm1cli.utils.generic.TM1Service", MockedTM1Service)
+    args = [raw_option] if raw_option else []
+    args += ["process", command]
+    result = runner.invoke(app, args)
+    assert result.exit_code == 0
+    assert isinstance(result.stdout, str)
+    assert result.stdout == expected_output
+
+
 @pytest.mark.parametrize(
     "raw_option,expected_output",
     [(None, "✅ Process exists!\n"), ("--output-raw", "True\n")],