+++ tm1cli/commands/subset.py
@@ -1,12 +1,9 @@
 from typing import Annotated
 
 import typer
-from rich import print  # pylint: disable=redefined-builtin
-from TM1py.Services import TM1Service
 
 from tm1cli.utils.cli_param import DATABASE_OPTION, INTERVAL_OPTION, WATCH_OPTION
-from tm1cli.utils.generic import execute_exists
-from tm1cli.utils.various import resolve_database
+from tm1cli.utils.generic import execute_exists, generic_list
 from tm1cli.utils.watch import watch_option
 
 app = typer.Typer()
@@ -24,9 +21,7 @@ def list_subset(
     List subsets
     """
 
-    with TM1Service(**resolve_database(ctx, database)) as tm1:
-        for subset in tm1.subsets.get_all_names(dimension_name):
-            print(subset)
+    generic_list("subsets", ctx, database, dimension_name=dimension_name)
 
 
 @app.command()